requirements.txt
Django==5.2.9
django-environ
django-cors-headers
strawberry-graphql[cli]
strawberry-graphql-django
django-tinymce
django-modeltranslation
django-filter
Pillow
django-cities-light
psycopg[binary]
requests
groq
fcm-django
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
